+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Surrogate` models now operate on dataframes in experimental representation instead of
   tensors in computational representation
 - `Surrogate.posterior` models now returns a `Posterior` object
+- `param_bounds_comp` of `SearchSpace`, `SubspaceDiscrete` and `SubspaceContinuous` has
+  been replaced with `comp_rep_bounds`, which returns a dataframe
 
 ### Added
 - `Surrogate` base class now exposes a `to_botorch` method
@@ -33,6 +35,10 @@ _ `_optional` subpackage for managing optional dependencies
 - `transform` methods of `SearchSpace`, `SubspaceDiscrete` and `SubspaceContinuous`
   now take additional `allow_missing` and `allow_extra` keyword arguments
 - `GaussianSurrogate` base class for surrogate models with Gaussian posteriors
+- `comp_rep_columns` property for `Parameter`, `SearchSpace`, `SubspaceDiscrete`
+  and `SubspaceContinuous` classes
+- Reworked mechanisms for surrogate input/output scaling configurable per class
+- `ParameterScalerProtocol` class for enabling user-defined input scaling mechanisms
 
 ### Changed
 - Passing an `Objective` to `Campaign` is now optional
@@ -44,6 +50,8 @@ _ `_optional` subpackage for managing optional dependencies
 - Context information required by `Surrogate` models is now cleanly encapsulated into
   a `context` object passed to `Surrogate._fit`
 - Fallback models created by `catch_constant_targets` are stored outside of surrogate
+- `to_tensor` now also handles `numpy` arrays
+- `GaussianProcessSurrogate` no longer uses a separate scaling approach
 
 ### Removed
 - Support for Python 3.9 removed due to new [BoTorch requirements](https://github.com/pytorch/botorch/pull/2293) 
@@ -64,8 +72,7 @@ _ `_optional` subpackage for managing optional dependencies
 - Passing a dataframe via the `data` argument to the `transform` methods of
   `SearchSpace`, `SubspaceDiscrete` and `SubspaceContinuous` is no longer possible.
   The dataframe must now be passed as positional argument.
-- Role of `register_custom_architecture` has been taken over by
-  `baybe.surrogates.base.SurrogateProtocol`
+- A deprecation error is thrown when attempting to use `register_custom_architecture`
 
 ## [0.9.1] - 2024-06-04
 ### Changed